requirements.txt
python-dotenv
requests
pandas
flickrapi
pillow